Renamed main.rb and split stream.rb into two separate files.
